moved the cofig to tests dir. changed jest.config to ignore this file
